Conflicts:

	app/controllers/dashboard/configuration_controller.rb
	app/controllers/dashboard_controller.rb
	app/views/dashboard/configuration/_room.html.erb
	app/views/dashboard/configuration/edit_tournament.html.erb
	app/views/layouts/dashboard.html.erb
	db/migrate_data/001_initial.rb
	test/fixtures/cards.yml
	test/unit/card_test.rb

+++ app/controllers/dashboard/entry_controller.rb
@@ -10,30 +10,44 @@ class Dashboard::EntryController < DashboardController
   		@last_round = @last_round.number
   	end
   	
-  	@games_to_enter = Game.find(:all, :conditions => ['play_complete = ? AND (entry_complete IS NULL OR entry_complete = ?) AND (ignore_indivs IS NULL OR ignore_indivs = ?)', 't','f','f'], :include => :team_games)
-  	@teams_to_enter = @games_to_enter.collect{|g| g.teams}.flatten.uniq.sort{|a,b| a.name <=> b.name}  	
+  	@rounds_to_enter = Round.find(:all, :conditions => ['rounds.entry_complete IS NULL OR rounds.entry_complete != ?', 't'], :include => :games)
+  	
   end
 
   def teams_for_round
-  	@teams = get_teams_for_round params[:id].to_i
-	render :partial => 'options_for_round', :collection => @teams  
+  	allt = params[:editing] ? true : false
+  	@teams = get_teams_for_round params[:id].to_i, allt
+	render :partial => 'options_for_round', :collection => @teams, :locals => {:eltsel => nil}
   end
   
-  def get_teams_for_round(round)
-    	Team.find(:all, :order => 'name').select{|t| t.games.select{|g| g.round.number == round }.empty? }
+  def get_teams_for_round(round, allt = false)
+  	if allt
+  		Team.find(:all, :order => 'name')
+  	else
+	    	Team.find(:all, :order => 'name').select{|t| t.games.select{|g| g.round.number == round }.empty? }
+	end
   end
   
   def rooms_for_round
-  	@rooms = get_rooms_for_round params[:id].to_i
-  	render :partial => 'options_for_round', :collection => @rooms
+  	allr = params[:editing] ? true : false
+  	@rooms = get_rooms_for_round params[:id].to_i, allr
+	render :partial => 'options_for_round', :collection => @rooms, :locals => {:eltsel => params[:selected].nil? ? false : params[:selected]}
   end
   
-  def get_rooms_for_round(round)
-    	Room.find(:all, :order => 'name').select{|r| r.games.select{|g| g.round.number == round }.empty? }
+  def get_rooms_for_round(round, allr = false)
+  	if allr
+  		Room.find(:all, :order => 'name')
+  	else
+    		Room.find(:all, :order => 'name').select{|r| r.games.select{|g| g.round.number == round }.empty? }
+    	end
   end
   
-  def get_cards_for_round(round)
-  	Card.find(:all, :order => 'number').select{|c| c.team_games.select{|tg| tg.game.round.number == round }.empty? }
+  def get_cards_for_round(round, allc = false)
+  	if allc
+  		Card.find(:all, :order => 'number')
+  	else
+  		Card.find(:all, :order => 'number').select{|c| c.team_games.select{|tg| tg.game.round.number == round }.empty? }
+  	end
   end
   
   def bracket_for_team
@@ -41,9 +55,15 @@ class Dashboard::EntryController < DashboardController
   end
   
   def save_game
+  	begin
+  		@game = Game.find(params[:id])
+  	rescue ActiveRecord::RecordNotFound
+  		@game = nil
+  	end
+  
   	if params[:round_number].empty?
   		flash[:error] = "Round number cannot be empty."
-  		redirect_to :action => 'index'
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
   		return false
   	end
   	
@@ -62,70 +82,99 @@ class Dashboard::EntryController < DashboardController
 
 	if params[:team1] == params[:team2]
 		flash[:error] = "You cannot select the same team for both slots."
-		redirect_to :action => 'index'
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
 		return false
 	end	
   	
   	if params[:score1].to_i == params[:score2].to_i and not params[:forfeit]
   		flash[:error] = "Scores cannot be equal in a non-forfeit."
-  		redirect_to :action => 'index'
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
   		return false
   	end
 
-	teams = get_teams_for_round(round.number)
-	rooms = get_rooms_for_round(round.number)
-	cards = get_cards_for_round(round.number)
+	teams = get_teams_for_round(round.number, !@game.nil?)
+	rooms = get_rooms_for_round(round.number, !@game.nil?)
+	cards = get_cards_for_round(round.number, !@game.nil?)
 	
 	if not params[:extragame] and not (teams.include? Team.find(params[:team1]) and teams.include? Team.find(params[:team2]))
 		flash[:error] = "One or both teams has already played a game this round."
-		redirect_to :action => 'index'
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
 		return false
 	end
 	
-	if @tournament.tracks_rooms  and not params[:extragame] and not rooms.include? Room.find(params[:room])
+	if @tournament.tracks_rooms?  and not params[:extragame] and not rooms.include? Room.find(params[:room])
 		flash[:error] = "That room has already been used during this round."
-		redirect_to :action => 'index'
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
 		return false
 	end
 	
-	if @tournament.swiss and not params[:extragame] and not (cards.include? Card.find(params[:card1]) and cards.include? Card.find(params[:card2]) )
+	if @tournament.swiss? and not params[:extragame] and not (cards.include? Card.find_by_number(params[:card1]) and cards.include? Card.find_by_number(params[:card2]) )
 		flash[:error] = "One or both cards has already been used during this round."
-		redirect_to :action => 'index'
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
 		return false
 	end
 	
-  	game = round.games.build(:round => round, :tossups => params[:tossups], :extragame => params[:extragame], :overtime => params[:overtime], :playoffs => params[:playoffs], :forfeit => params[:forfeit])
-  	game.room = Room.find(params[:room]) if @tournament.tracks_rooms
-  	game.bracket = Bracket.find(params[:bracket]) if @tournament.bracketed
+	if @game.nil?
+	  	game = round.games.build(:round => round, :tossups => params[:tossups], :extragame => params[:extragame], :overtime => params[:overtime], :playoffs => params[:playoffs], :forfeit => params[:forfeit], :play_complete => true)
+  		game.room = Room.find(params[:room]) if @tournament.tracks_rooms?
+	  	game.bracket = Bracket.find(params[:bracket]) if @tournament.bracketed?
+	else
+		game = @game
+		game.update_attributes(:round => round, :tossups => params[:tossups], :extragame => params[:extragame], :overtime => params[:overtime], :playoffs => params[:playoffs], :forfeit => params[:forfeit], :play_complete => true)
+		game.room = Room.find(params[:room]) if @tournament.tracks_rooms?
+		game.bracket = Bracket.find(params[:bracket]) if @tournament.bracketed?
+	end
     	
     	if not game.save
     		game.errors.each_full {|msg| flash[:error] = msg }
-    		redirect_to :action => "index"
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
     		return false
     	end
     	
-	tg1 = game.team_games.build(:team => Team.find(params[:team1]), :points => params[:score1])
-    	tg1.card = Card.find(params[:card1]) if @tournament.swiss
+    	if @game.nil?
+		tg1 = game.team_games.build(:team => Team.find(params[:team1]), :points => params[:score1], :ordering => 1)
+	    	tg1.card = Card.find_by_number(params[:card1]) if @tournament.swiss?
     	
-    	tg2 = game.team_games.build(:team => Team.find(params[:team2]), :points => params[:score2])
-    	tg2.card = Card.find(params[:card2]) if @tournament.swiss
+    		tg2 = game.team_games.build(:team => Team.find(params[:team2]), :points => params[:score2], :ordering => 2)
+	    	tg2.card = Card.find_by_number(params[:card2]) if @tournament.swiss?
+	else
+		tg1 = game.team_games[0]
+		tg1.update_attributes(:points => params[:score1])
+		tg1.card = Card.find_by_number(params[:card1]) if @tournament.swiss?
+		
+		tg2 = game.team_games[1]
+		tg2.update_attributes(:points => params[:score2])
+		tg2.card = Card.find_by_number(params[:card2]) if @tournament.swiss?
+	end
+	
+	if tg1.points > tg2.points
+		tg1.won = true
+		tg2.won = false
+	else
+		tg1.won = false
+		tg2.won = true
+	end
     		
     	if not tg1.save
     		tg1.errors.each_full {|msg| flash[:error] = msg }
-    		game.destroy
-    		redirect_to :action => "index"
+    		game.destroy if @game.nil?
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
     		return false
     	end
     		
     	if not tg2.save
     		tg1.errors.each_full {|msg| flash[:error] = msg }
-    		game.destroy
-    		redirect_to :action => "index"
+    		game.destroy if @game.nil?
+		redirect_to @game.nil? ? {:action => 'index'} : {:action => 'edit_game', :id => @game.id}
     		return false
     	end
     		
+    	expire_page :controller => 'statistics', :action => 'standings'
+    	expire_page :controller => 'statistics', :action => 'team', :id => tg1.team.id
+    	expire_page :controller => 'statistics', :action => 'team', :id => tg2.team.id
+    	expire_page :controller => 'statistics', :action => 'scoreboard'
     	flash[:notice] = "Game saved."
-    	redirect_to :action => 'index'
+	redirect_to @game.nil? ? {:action => 'index'} : {:action => 'status'}
   end
   
   def ignore_indivs
@@ -149,14 +198,12 @@ class Dashboard::EntryController < DashboardController
   
   def enter_indivs
   	begin
-	  	@teamgame1 = TeamGame.find(params[:id])
+	  	@game = Game.find(params[:id])
 	rescue ActiveRecord::RecordNotFound
 		flash[:error] = "The requested game was not found."
 		redirect_to :action => 'index'
 		return false
 	end
-	
-	@game = @teamgame1.game
 	
 	if @game.ignore_indivs
 		flash[:error] = "The selected game has been marked as ignored."
@@ -164,7 +211,8 @@ class Dashboard::EntryController < DashboardController
 		return false
 	end
 	
-	@teamgame2 = @game.team_game_for_other(@teamgame1.team)
+	@teamgame1 = @game.team_games[0]
+	@teamgame2 = @game.team_games[1]
 	@team1 = @teamgame1.team
 	@team2 = @teamgame2.team
 	@types = QuestionType.find(:all, :order => 'value desc')
@@ -193,7 +241,7 @@ class Dashboard::EntryController < DashboardController
         		if pgame.tossups_heard > game.tossups
         			#fail
         			flash[:error] = "Player tossups heard were greater than game tossups heard."
-        			redirect_to :action => 'enter_indivs', :id => game.team_games.first.id
+        			redirect_to :action => 'enter_indivs', :id => game.id
         			pgame.destroy
         			pgs.each{|pg| pg.destroy}
         			return false;
@@ -206,7 +254,7 @@ class Dashboard::EntryController < DashboardController
         		if pgame.stat_lines.collect{|sl| sl.number}.sum > pgame.tossups_heard
         			#fail
         			flash[:error] = "Player answered more tossups than tossups heard."
-        			redirect_to :action => 'enter_indivs', :id => game.team_games.first.id
+        			redirect_to :action => 'enter_indivs', :id => game.id
         			pgame.destroy
         			pgs.each{|pg| pg.destroy}
         			return false;        			
@@ -244,7 +292,7 @@ class Dashboard::EntryController < DashboardController
 	if tot_tossups.values.sum > game.tossups
 		# fail		
 		flash[:error] = "More tossups were answered correctly than were asked."
-		redirect_to :action => 'enter_indivs', :id => game.team_games.first.id
+		redirect_to :action => 'enter_indivs', :id => game.id
 		pgs.each{|pg| pg.destroy}
 		return false;
 	end
@@ -252,7 +300,7 @@ class Dashboard::EntryController < DashboardController
 	if tot_tuh > (8*game.tossups)
 		#fail
 		flash[:error] = "More tossups were heard by all players than the maximum possible."
-		redirect_to :action => 'enter_indivs', :id => game.team_games.first.id
+		redirect_to :action => 'enter_indivs', :id => game.id
 		pgs.each{|pg| pg.destroy}
 		return false;
 	end
@@ -261,7 +309,7 @@ class Dashboard::EntryController < DashboardController
 		if tot_ans[team.id] > game.tossups
 			#fail
 			flash[:error] = "More tossups were answered by the team than were asked."
-			redirect_to :action => 'enter_indivs', :id => game.team_games.first.id
+			redirect_to :action => 'enter_indivs', :id => game.id
 			pgs.each{|pg| pg.destroy}
 			return false;
 		end
@@ -269,7 +317,7 @@ class Dashboard::EntryController < DashboardController
 		if tot_tossups[team.id] == 0 and bps[team.id] > 0
 			#fail			
 			flash[:error] = "Team has bonus points without any correct tossups."
-			redirect_to :action => 'enter_indivs', :id => game.team_games.first.id
+			redirect_to :action => 'enter_indivs', :id => game.id
 			pgs.each{|pg| pg.destroy}
 			return false;
 		elsif tot_tossups[team.id] > 0
@@ -277,7 +325,7 @@ class Dashboard::EntryController < DashboardController
 			if bppt < 0.0 or bppt > 30.0
 				#fail				
 				flash[:error] = "Bonus points per tossup correct is out of range 0-30"
-				redirect_to :action => 'enter_indivs', :id => game.team_games.first.id
+				redirect_to :action => 'enter_indivs', :id => game.id
 				pgs.each{|pg| pg.destroy}
 				return false;
 			end
@@ -293,8 +341,66 @@ class Dashboard::EntryController < DashboardController
 	    	
 	game.entry_complete = true
 	game.save
+	expire_page :controller => 'statistics', :action => 'personal'
 	flash[:notice] = "Individual standings saved."
 	redirect_to :action => 'index'  
+  end
+  
+  # Will list all entered games and link for editing
+  def status
+  	@rounds = Round.find(:all, :include => :games, :order => 'number')
+  	
+  end
+  
+  def edit_game
+  	begin
+  		@game = Game.find(params[:id])
+  	rescue ActiveRecord::RecordNotFound
+  		flash[:error] = "Game was not found."
+  		redirect_to :action => 'status'
+  	end
+  
+  	@last_round = @game.round.number
+  end
+  
+  def delete_game
+  	begin
+  		@game = Game.find(params[:id])
+  	rescue ActiveRecord::RecordNotFound
+  		flash[:error] = "Game was not found."
+  		redirect_to :action => 'status'
+  	end
+  	
+  	@game.destroy
+  	
+  	expire_page :controller => 'statistics', :action => 'standings'
+    	expire_page :controller => 'statistics', :action => 'team', :id => tg1.team.id
+    	expire_page :controller => 'statistics', :action => 'team', :id => tg2.team.id
+    	expire_page :controller => 'statistics', :action => 'scoreboard'
+  	flash[:notice] = "Game was deleted successfully."
+  	redirect_to :action => 'status'
+  end
+  
+  def clear_indivs
+  	begin
+  		@game = Game.find(params[:id])
+  	rescue ActiveRecord::RecordNotFound
+  		flash[:error] = "Game was not found."
+  		redirect_to :action => 'status'
+  	end
+  	
+  	@game.team_games.each do |tg|
+  		tg.player_games.each do |pg|
+  			pg.destroy
+  		end
+  	end
+  	
+  	@game.update_attributes :entry_complete => false
+  	
+  	expire_page :controller => 'statistics', :action => 'personal'
+  	flash[:notice] = "Individual stats were cleared for that game."
+  	redirect_to :action => 'status'
+  
   end
   
 end
